+++ rust/operator-binary/src/spark_k8s_controller/validate.rs
@@ -16,7 +16,10 @@ use stackable_operator::{
 };
 
 use crate::{
-    crd::{constants::CONTAINER_IMAGE_BASE_NAME, logdir::ResolvedLogDir, v1alpha1},
+    crd::{
+        ResolvedSparkApplicationTemplate, constants::CONTAINER_IMAGE_BASE_NAME,
+        logdir::ResolvedLogDir, v1alpha2,
+    },
     spark_k8s_controller::dereference::DereferencedSparkApplication,
 };
 
@@ -38,8 +41,8 @@ type Result<T, E = Error> = std::result::Result<T, E>;
 
 /// Inputs the rest of `reconcile` needs after dereferencing.
 pub struct ValidatedSparkApplication {
-    pub spark_application: v1alpha1::SparkApplication,
-    pub resolved_template_refs: Vec<v1alpha1::ResolvedSparkApplicationTemplate>,
+    pub spark_application: v1alpha2::SparkApplication,
+    pub resolved_template_refs: Vec<ResolvedSparkApplicationTemplate>,
     pub s3_connection: Option<s3::v1alpha1::ConnectionSpec>,
     pub log_dir: Option<ResolvedLogDir>,
     pub resolved_product_image: ResolvedProductImage,